+++ src/main/java/org/betonquest/betonquest/compatibility/citizens/NPCLocationCondition.java
@@ -10,7 +10,6 @@ import org.betonquest.betonquest.exceptions.InstructionParseException;
 import org.betonquest.betonquest.exceptions.QuestRuntimeException;
 import org.betonquest.betonquest.utils.location.CompoundLocation;
 import org.bukkit.Location;
-import org.bukkit.entity.Entity;
 
 /**
  * Checks if a npc is at a specific location
@@ -43,15 +42,11 @@ public class NPCLocationCondition extends Condition {
         if (npc == null) {
             throw new QuestRuntimeException("NPC with ID " + npcId + " does not exist");
         }
-        final Entity npcEntity = npc.getEntity();
-        if (npcEntity == null) {
-            return false;
-        }
         final Location location = this.location.getLocation(playerID);
-        if (!location.getWorld().equals(npcEntity.getWorld())) {
+        if (!location.getWorld().equals(npc.getStoredLocation().getWorld())) {
             return false;
         }
         final double radius = this.radius.getDouble(playerID);
-        return npcEntity.getLocation().distanceSquared(location) <= radius * radius;
+        return npc.getStoredLocation().distanceSquared(location) <= radius * radius;
     }
 }