+++ src/main/java/org/betonquest/betonquest/compatibility/citizens/NPCRangeObjective.java
@@ -11,7 +11,6 @@ import org.betonquest.betonquest.exceptions.QuestRuntimeException;
 import org.betonquest.betonquest.utils.PlayerConverter;
 import org.bukkit.Bukkit;
 import org.bukkit.Location;
-import org.bukkit.entity.Entity;
 import org.bukkit.entity.Player;
 
 import java.util.*;
@@ -65,12 +64,8 @@ public class NPCRangeObjective extends Objective {
             if (npc == null) {
                 throw new QuestRuntimeException("NPC with ID " + npcId + " does not exist");
             }
-            final Entity npcEntity = npc.getEntity();
-            if (npcEntity == null) {
-                return;
-            }
             for (final Player player : Bukkit.getOnlinePlayers()) {
-                if (!playersInside.contains(player.getUniqueId()) && isInside(player, npcEntity.getLocation())) {
+                if (!playersInside.contains(player.getUniqueId()) && isInside(player, npc.getStoredLocation())) {
                     playersInside.add(player.getUniqueId());
                 }
             }